test-requirements.txt
hacking>=0.9.2,<0.10

coverage>=3.6
discover
fixtures>=0.3.14
python-subunit>=0.0.18
sphinx>=1.1.2,!=1.2.0,!=1.3b1,<1.3
oslosphinx>=2.2.0
oslotest>=1.2.0  # Apache-2.0
testrepository>=0.0.18
testscenarios>=0.4
testtools>=0.9.36,!=1.2.0
mock>=1.0
mox>=0.5.3
